requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
pytest
httpx
pydantic
fastapi
uvicorn[standard]
google-generativeai
fastapi-pagination